Split out integrations, add code section comments
